requirements.txt
aiml==0.9.2
aiohttp==3.6.2
asn1crypto==1.2.0
astroid==2.3.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
idna==2.8
isort==4.3.21
joblib==0.13.2
lazy-object-proxy==1.4.2
mccabe==0.6.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
multidict==4.5.2
numpy==1.17.2
olefile==0.46
Pillow==6.2.0
pycparser==2.19
pylint==2.4.3
pyOpenSSL==19.0.0
PySocks==1.7.1
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
scrython==1.7.1
six==1.12.0
urllib3==1.24.2
wrapt==1.11.2
yarl==1.3.0
